Changed tag detail advanced editing layout
ref https://linear.app/ghost/issue/PLA-342/design-refinements

Closing a code injection modal made it unclear whether edits were saved. Keeping injection inline beneath core data preserves the page-level save model, while an always-visible metadata card removes unnecessary disclosure in the right rail.
